pyproject.toml
[project]
name = "assignment-1-fastapi-alembic-sqlmodel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.134.0",
    "pydantic-settings>=2.13.1",
    "scalar-fastapi>=1.6.2",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.41.0",
]
